[#241] ファイルが存在しない場合のニックネーム変更でエラーが起こらない様に
git-svn-id: http://svn.openskip.org/skip/trunk@418 433210a6-3911-11dd-bd70-bf985eec0d0c

+++ app/models/admin/user_uid.rb
@@ -38,7 +38,7 @@ class Admin::UserUid < UserUid
       ActiveRecord::Base.connection.execute("update mails set to_address_symbol = 'uid:#{uid}' where to_address_symbol = 'uid:#{uid_was}'")
       ActiveRecord::Base.connection.execute("update bookmarks set url = replace(url, '/user/#{uid_was}', '/user/#{uid}') where url = '/user/#{uid_was}'")
       ActiveRecord::Base.connection.execute("update antenna_items set value = 'uid:#{uid}' where value = 'uid:#{uid_was}'")
-      FileUtils.mv(ShareFile.dir_path("uid:#{uid_was}"), ShareFile.dir_path("uid:#{uid}"))
+      FileUtils.mv(ShareFile.dir_path("uid:#{uid_was}"), ShareFile.dir_path("uid:#{uid}")) if File.exist?(ShareFile.dir_path("uid:#{uid_was}"))
     end
   end
